% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_function.R
\name{a_function}
\alias{a_function}
\title{A dummy function}
\usage{
a_function(X)
}
\arguments{
\item{X}{dummy variable}
}
\value{
whatever dummy variable is provided.
}
\description{
Use for testing Allows incorporation of new sample data (a dataframe) to be merged into an existing phyloseq object.
}
