% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assignment}
\alias{assignment}
\title{Example ASV assignments from assignSpecies_mod}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
assignment
}
\description{
\describe{
\item{asv}{Amplicon sequence variant (ASV) identified}
\item{index}{The NCBI accession number of the sequence's exact match in the
reference database}
\item{label}{The taxonomic name associated with the NCBI accession number}
}
}
\keyword{datasets}
