% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSpecies_mod.R
\name{assignSpecies_mod}
\alias{assignSpecies_mod}
\title{Return long-form species assignments with DADA2's assignSpecies}
\usage{
assignSpecies_mod(seqs, refFasta, tryRC = FALSE, n = 2000)
}
\arguments{
\item{seqs}{A character vector of the sequences to be assigned}

\item{refFasta}{The path to the reference fasta file, or an R connection. Can
  be compressed. This reference fasta file should be formatted so that the ID
  lines correspond to the genus-species of the associated sequence:

  >SeqID genus species ACGAATGTGAAGTAA......}

\item{tryRC}{(Optional). Default FALSE. If TRUE, the reverse-complement of
each sequences will also be tested for exact matching to the reference
sequences.}

\item{n}{(Optional). Default 2000. The number of sequences to perform
assignment on at one time. This controls the peak memory requirement so
that large numbers of sequences are supported.}
}
\value{
A two-column data frame matching ASVs to their exact match in the
  reference, with multiple matches indicated by the presence of more than one
  row per ASV.
}
\description{
This function is a modification of DADA2's \code{assignSpecies},
  which uses exact matching against a reference fasta to identify the
  genus-species binomial classification of the input sequences. Where the
  original function run with \code{allowMultiple = TRUE} returns a
  concatenated string of all exactly matched species, results are returned
  here as distinct rows, one per match.
}
