% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_to_taxonomy.R
\name{asv_to_taxonomy}
\alias{asv_to_taxonomy}
\title{Join ASV assignments to taxonomy}
\usage{
asv_to_taxonomy(assignment, taxonomy)
}
\arguments{
\item{assignment}{The output of \code{assignSpecies_mod}}

\item{taxonomy}{An \code{R6Class} object of class \code{Taxonomy} produced by
the \code{metacoder::taxonomy} function.}
}
\value{
A join of each ASV assignment to its full taxonomy, done from finer
  taxonomic levels upward.
}
\description{
The output of the \code{assignSpecies_mod} function is a
  two-column dataframe of ASVs and their taxonomic assignments based on exact
  matching to reference sequences.  This joins the named taxonomic assignment
  to a full taxonomy queried from NCBI, assuming that all assignments occur
  to at least the species level, and potentially to subspecies levels
  (currently hard-coded for subspecies, variety, and form).
}
