% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{axes}
\alias{axes}
\title{Bootstrap-branded axes and labels for base R graphics}
\usage{
axes(
  side = getOption("axes.side"),
  at = list(x = NULL, y = NULL),
  axes.lty = NULL,
  gap.axis = c(NA, NA),
  line = c(0, 0),
  main = NA,
  sub = NA,
  xlab = NA,
  ylab = NA,
  nx = NA,
  ny = NULL,
  grid.col = par("col"),
  grid.lty = "dotted",
  grid.lwd = par("lwd")
)
}
\arguments{
\item{side}{position of axes, up to 4 (1-bottom, 2-left, 3-top, 4-right)}

\item{at}{length-2 list of x/y points at which tickmarks are to be drawn, default \code{list(x=NA, y=NA)}}

\item{axes.lty}{length-2 type of axis line (depends on \code{nx} value, c(0,1) if vertical gridlines are drawn) but can be modified here}

\item{gap.axis}{length-2 minimal gap between axis labels}

\item{line}{specifying a value for \code{line} overrides the
    default placement of labels, and places them this many lines
    outwards from the plot edge.}

\item{main}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}

\item{sub}{Sub-title (at bottom) using font, size and color 
    \code{par(c("font.sub", "cex.sub", "col.sub"))}.}

\item{xlab}{X axis label using font, size and color 
    \code{par(c("font.lab", "cex.lab", "col.lab"))}.}

\item{ylab}{Y axis label, same font attributes as \code{xlab}.}

\item{nx, ny}{number of cells of the grid in x and y direction.  When
    \code{NULL}, as per default, the grid aligns with the tick marks on
    the corresponding \emph{default} axis (i.e., tickmarks as computed by
    \code{\link[graphics]{axTicks}}).  When \code{\link{NA}}, no grid lines are
    drawn in the corresponding direction.}

\item{grid.col}{gridline color}

\item{grid.lty}{gridline type}

\item{grid.lwd}{gridline width}
}
\value{
Branded axis and grid for base R plots
}
\description{
Shorthand function to apply Bootstrap brand to base R plot axes and labels. Many tips derived from \href{https://r-charts.com/base-r/margins/}{r-charts.com}, since base R graphic documentation is often lacking.
}
\details{
This function has no side effect and does not modify the current device, but it does require at minimum to specifiy \code{par(mar = c(2, 2, 7, 3))} to provide margin space for all plot labels.
}
\examples{
set.seed(1)
x = runif(100, min = -5, max = 5)
y = x ^ 3 + rnorm(100, mean = 0, sd = 5)

opar = par(par.brand())

plot(x, y)
axes(nx=NULL, xlab="X Units", ylab="Y Units",
  main="My Bootstrap Branded Plot", sub="Subtitle")
abline(h=0, col="red", lwd=2)

plot(x, type="h", col=c("red", "green")[(x > 0) + 1])
axes(xlab="X units", ylab="Y units",
  main="My Bootstrap Branded Plot", sub="Histogram")
legend_brand(legend=c("Red", "Green"), lty=1, lwd=2, col=c("red", "green"))

hist(x, col=pal(), border=NA)
axes(c(1,4), xlab=NA, ylab="Frequency",
  main="My Bootstrap Branded Plot", sub="Histogram, dummy legend")
legend_brand(legend=paste("cat", 1:3), fill=pal(1:3), lty=0)

# Restore
par(opar)

}
