% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot}
\alias{biplot}
\title{Flexibly generate a biplot from the output of a PCA}
\usage{
biplot(pca, top.n, data, color, base_only)
}
\arguments{
\item{pca}{The output of `prcomp()` on the data to be ordinated.}

\item{top.n}{The number of loadings to plot (ranked by descending magnitude).
Default 5}

\item{data}{A `data.frame` of sample metadata variables row names matching
those in the PCA output.}

\item{color}{The sample metadata variable (a column in provided `data`) by
which to color plotted points.  Taken as a string (i.e. 'subj')}

\item{base_only}{Should only the base PCA be returned, without loadings?
Default FALSE}
}
\value{
A ggplot object
}
\description{
Given the output of a PCA, makes biplot of data showing the
  magnitude of the factor loadings that contribute the plotted PCs.
}
