% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\docType{data}
\name{BRAND}
\alias{BRAND}
\title{Built-in default Bootstrap brand}
\format{
List of default branding elements (colors, fonts)
}
\usage{
BRAND
}
\description{
Branding is read from a user-specified \verb{_brand.yml} file or from this package if
no file is found in the working tree.
}
\examples{
names(BRAND)

}
\references{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
\keyword{dataset}
