% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{brand_off}
\alias{brand_off}
\title{Turn Bootstrap brand off}
\usage{
brand_off()
}
\value{
TRUE (invisibly)
}
\description{
Restore R session to its original state, discarding all color, font and plot
customizations.
}
\seealso{
\code{\link[=brand_on]{brand_on()}}
}
