% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder_internal.R
\name{check_for_pkg}
\alias{check_for_pkg}
\title{check for packages}
\usage{
check_for_pkg(package)
}
\arguments{
\item{package}{The name of the package}
}
\value{
`TRUE` if package is present
}
\description{
check for packages, and stop if not installed
}
\keyword{internal}
