% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.annotation.R
\name{date.annotation}
\alias{date.annotation}
\title{Copy directory to R understanding format.}
\usage{
date.annotation(date = Sys.time())
}
\arguments{
\item{date}{"POSIXct" "POSIXt" file. Default Sys.time()}
}
\value{
"character" traslation of date time as YYYYMMDD_HHMMSS
}
\description{
Copy directory to R understanding format.
}
\examples{
date.annotation(date = "2022-12-14 13:09:53 CET")
##"20221214_130953"
date.annotation()
##"20221214_130953"
}
\author{
Manuel Molina-Bustamante
}
