% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endingtime.R
\name{endingtime}
\alias{endingtime}
\title{Calculate the Ending Time of a Process}
\usage{
endingtime(start.time, duration)
}
\arguments{
\item{start.time}{A character string representing the start time in "hh:mm"
format.}

\item{duration}{A character string representing the duration in "hh:mm"
format.}
}
\value{
A POSIXct object representing the end time in "yyyy-mm-dd hh:mm:ss"
format.
}
\description{
This function calculates the ending time of a process given a start time and
a duration.
}
\examples{
endingtime("08:00", "03:30")

}
