% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggbrand}
\alias{ggbrand}
\title{Bootstrap-branded ggplot with custom axis and legend}
\usage{
ggbrand(
  data = NULL,
  mapping = aes(),
  axes = c("bottomright", "topright", "bottomleft", "topleft"),
  ...
)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{axes}{position of X (\code{bottom} or \code{top}) and Y (\code{left} or \code{right}) primary axes (default: \code{bottomright})}

\item{...}{
  Arguments passed on to \code{\link{theme_brand}}
  \describe{
    \item{\code{base_color}}{Bootstrap color name for text and line elements, or color code}
    \item{\code{base_bg}}{Bootstrap color name for plot, panel background, or color code}
    \item{\code{base_family}}{one of \verb{_brand.yml} font families (currently only \code{base}, \code{monospace}, or \code{headings}), else a valid system or Google font family name}
    \item{\code{grid}}{show gridlines \code{XY}, \code{X}, \code{Y} (default) or \code{n} for no gridline}
    \item{\code{legend}}{shorthand for \code{theme(legend.position="...")} (default: \code{top})}
    \item{\code{gradient}}{names of Bootstrap colors for continuous gradient (default: \code{c("blue", "orange", "red")})}
    \item{\code{base_size}}{base font size, given in pts.}
  }}
}
\value{
A \code{ggplot2} object with themef elements
}
\description{
Shorthand to generate branded ggplots. Uses \code{\link[=theme_brand]{theme_brand()}} to control colors and fonts, and has Y-axis on the right. Axis placement can be quickly modified with argument \code{axes}.
}
\examples{
require(ggplot2)

ggbrand(mtcars, aes(factor(carb), mpg, fill=factor(carb))) +
  geom_col() +
  labs(
    x = "carb",
    title = "Default Plot with Y-axis on the Right",
    subtitle = "My very long subtitle with many units",
    caption = "My very long plot caption with many references.")

# Equivalent to below
ggplot(mtcars, aes(factor(carb), mpg, fill=factor(carb))) +
  geom_col() +
  scale_brand_df() +
  guides(y=guide_axis(position="right")) +
  theme_brand(grid="XY")

ggbrand(mtcars, aes(wt, mpg, color=carb), axes="topright") +
  geom_smooth(color=pal("red"), fill=pal("pink")) +
  geom_point(size=3) +
  guides(color=guide_legend(nrow=1)) +
  labs(
    title = "My Beautiful Plot with X-axis at the Top",
    subtitle = "My descriptive subtitle with units",
    caption = "My plot caption with many references.")

ggbrand(mtcars, aes(wt, mpg, color=factor(cyl)),
  grid="XY", legend="bottom", base_family="Pacifico") +
  geom_point(size=3) +
  labs(
    x = "cyl",
    title = "My Plot with Full Gridlines and Pacifico Font",
    subtitle = "Placed the legend at the bottom",
    caption = "My plot caption with many references.")

}
