% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{legend_brand}
\alias{legend_brand}
\title{Wrapper for graphics::legend()}
\usage{
legend_brand(
  legend,
  x = "topright",
  y = NULL,
  bty = "n",
  horiz = TRUE,
  xpd = TRUE,
  inset = c(-0.05, -0.4),
  par = par.brand(),
  ...
)
}
\arguments{
\item{legend}{a character or \link{expression} vector
    of length \eqn{\ge 1} to appear in the legend.  Other
    objects will be coerced by \code{\link{as.graphicsAnnot}}.}

\item{x, y}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}

\item{bty}{the type of box to be drawn around the legend.  The allowed
    values are \code{"o"} (the default) and \code{"n"}.}

\item{horiz}{logical; if \code{TRUE}, set the legend horizontally
    rather than vertically (specifying \code{horiz} overrides the \code{ncol}
    specification).}

\item{xpd}{if supplied, a value of the \link[graphics]{graphical parameter} \code{xpd}
    to be used while the legend is being drawn.}

\item{inset}{inset distance(s) from the margins as a fraction of the
    plot region when legend is placed by keyword.}

\item{...}{
  Arguments passed on to \code{\link[graphics:legend]{graphics::legend}}
  \describe{
    \item{\code{x,y}}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}
    \item{\code{fill}}{if specified, this argument will cause boxes filled
    with the specified colors (or shaded in the specified colors)
    to appear beside the legend text.}
    \item{\code{col}}{the color of points or lines appearing in the legend.}
    \item{\code{border}}{the border color for the boxes (used only if
    \code{fill} is specified).}
    \item{\code{lty,lwd}}{the line types and widths for lines appearing in the
    legend.  One of these two \emph{must} be specified for line drawing.}
    \item{\code{pch}}{the plotting symbols appearing in the legend, as
    numeric vector or a vector of 1-character strings (see
    \code{\link[graphics]{points}}). Unlike \code{points}, this can all be
    specified as a single multi-character string.  \emph{Must} be
    specified for symbol drawing.}
    \item{\code{angle}}{angle of shading lines.}
    \item{\code{density}}{the density of shading lines, if numeric and
    positive. If \code{NULL} or negative or \code{NA} color filling
    is assumed.}
    \item{\code{bg}}{the background color for the legend box.  (Note that this is
    only used if \code{bty != "n"}.)}
    \item{\code{box.lty,box.lwd,box.col}}{the line type, width and color for
    the legend box (if \code{bty = "o"}).}
    \item{\code{pt.bg}}{the background color for the \code{\link[graphics]{points}},
    corresponding to its argument \code{bg}.}
    \item{\code{cex}}{character expansion factor \bold{relative} to current
    \code{par("cex")}.  Used for text, and provides the default for
    \code{pt.cex}.}
    \item{\code{pt.cex}}{expansion factor(s) for the points.}
    \item{\code{pt.lwd}}{line width for the points, defaults to the one for
    lines, or if that is not set, to \code{par("lwd")}.}
    \item{\code{xjust}}{how the legend is to be justified relative to the legend
    x location.  A value of 0 means left justified, 0.5 means centered
    and 1 means right justified.}
    \item{\code{yjust}}{the same as \code{xjust} for the legend y location.}
    \item{\code{x.intersp}}{character interspacing factor for horizontal (x) spacing
    between symbol and legend text.}
    \item{\code{y.intersp}}{vertical (y) distances (in lines of text shared above/below each legend entry).
    A vector with one element for each row of the legend can be used.}
    \item{\code{adj}}{numeric of length 1 or 2; the string adjustment for legend
    text.  Useful for y-adjustment when \code{labels} are
    \link{plotmath} expressions.}
    \item{\code{text.width}}{the width of the legend text in x (\code{"user"})
    coordinates.  (Should be positive even for a reversed x axis.)
    Can be a single positive numeric value (same width for each column of the legend),
    a vector (one element for each column of the legend),
    \code{NULL} (default) for computing a proper maximum value of \code{\link[graphics]{strwidth}(legend)}),
    or \code{NA} for computing a proper column wise maximum value of \code{\link[graphics]{strwidth}(legend)}).}
    \item{\code{text.col}}{the color used for the legend text.}
    \item{\code{text.font}}{the font used for the legend text, see \code{\link[graphics]{text}}.}
    \item{\code{merge}}{logical; if \code{TRUE}, merge points and lines but
    not filled boxes.  Defaults to \code{TRUE} if there are points and lines.}
    \item{\code{trace}}{logical; if \code{TRUE}, shows how \code{legend} does all
    its magical computations.}
    \item{\code{plot}}{logical.  If \code{FALSE}, nothing is plotted but the
    sizes are returned.}
    \item{\code{ncol}}{the number of columns in which to set the legend items
    (default is 1, a vertical legend).}
    \item{\code{title}}{a character string or length-one expression giving a
    title to be placed at the top of the legend.  Other objects will be
    coerced by \code{\link{as.graphicsAnnot}}.}
    \item{\code{title.col}}{color for \code{title}, defaults to \code{text.col[1]}.}
    \item{\code{title.adj}}{horizontal adjustment for \code{title}: see the help
    for \code{\link[graphics]{par}("adj")}.}
    \item{\code{title.cex}}{expansion factor(s) for the title, defaults to \code{cex[1]}.}
    \item{\code{title.font}}{the font used for the legend title, defaults to \code{text.font[1]},
    see \code{\link[graphics]{text}}.}
    \item{\code{seg.len}}{the length of lines drawn to illustrate \code{lty}
    and/or \code{lwd} (in units of character widths).}
  }}
}
\value{
Branded legend for base R plots
}
\description{
Simply assign default arguments to base R \code{\link[=legend]{legend()}}. This function is exported but is not meant to be called directly, use \code{\link[=brand_on]{brand_on()}} instead.
}
\examples{
set.seed(1)
x <- runif(100, min = -5, max = 5)
y <- x ^ 3 + rnorm(100, mean = 0, sd = 5)

opar <- par(par.brand())

plot(x, y, type="h", col=(y>0)+4)
axes(nx=NULL,
  main="Bootstrap Branded Plot", sub="My Subtitle",
  xlab="X Units", ylab="Y Units")
abline(h=0, col=pal("red"), lwd=2)
legend_brand(names(pal())[4:5], lty=1, lwd=2, col=4:5)

plot(x, type="h", col=pal())
axes(side=c(1,4),
  main="My Bootstrap Branded Plot",
  sub="Histogram, dummy legend", ylab="Frequency")
legend_brand(paste("cat", 1:3), fill=pal(1:3))

par(opar)

}
\seealso{
\code{\link[=axes]{axes()}}
}
