% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{par.brand}
\alias{par.brand}
\title{Set or query graphical parameters for Bootstrap branded plots}
\usage{
par.brand(fg = NULL, bg = NULL, family = NULL, ...)
}
\arguments{
\item{fg}{Bootstrap color \strong{name} for text and line elements, or color code}

\item{bg}{Bootstrap color \strong{name} for plot, panel background, or color code}

\item{family}{one of \verb{_brand.yml} font families (currently only \code{base}, \code{monospace}, or \code{headings}), else a valid System or Google font family name}

\item{...}{
  Arguments passed on to \code{\link[graphics:par]{graphics::par}}
  \describe{
    \item{\code{no.readonly}}{logical; if \code{TRUE} and there are no other
    arguments, only parameters are returned which can be set by a
    subsequent \code{par()} call \emph{on the same device}.}
  }}
}
\value{
a list of graphical parameters
}
\description{
Opinionated mods to base R plots compatible with new \verb{_brand.yml} feature. Does not load any brand by itself, explicitely call \code{\link[=brand_on]{brand_on()}} if needed instead.
}
\examples{
opar <- par(par.brand())
par()
# Restore device to default state
par(opar)

}
