% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_counts.R
\name{parse_counts}
\alias{parse_counts}
\title{Parse read count QC files generated at pipeline steps}
\usage{
parse_counts(f)
}
\arguments{
\item{f}{The path to a text file that alternates fastq.gz filenames with
their read counts.}
}
\value{
A two-column data frame matching samples to their read counts.
}
\description{
This function parses text files of sample read counts of
  generated during the metabarcoding analysis pipeline for QC purposes.
}
