% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_unassigned.R
\name{percent_unassigned}
\alias{percent_unassigned}
\title{Find the percentage of reads without a taxon assignment}
\usage{
percent_unassigned(ps, by_sample = FALSE)
}
\arguments{
\item{ps}{The phyloseq object to operate on.}

\item{by_sample}{(Default: FALSE) Whether or not to return the calculations
for the whole object, or broken out by individual samples in the dataset.}
}
\value{
The percentage of unmapped reads, either for the overall dataset or
  on a per-sample basis.
}
\description{
Identifies sequence variants with no taxon assignment (*i.e.*,
  those for which every column of the taxonomy table is NA).  Then,
  determines what percentage of overall reads in the dataset correspond to
  unidentified reads.  This is helpful for determining if the current
  reference database for taxonomic assignment is a good fit to the data.
}
