% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brand}
\alias{plot.brand}
\title{Wrapper for base plot()}
\usage{
\method{plot}{brand}(
  ...,
  frame.plot = FALSE,
  side = getOption("axes.side"),
  use.brand = getOption("use.brand")
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=axes]{axes}}, \code{\link[graphics:plot.default]{graphics::plot}}
  \describe{
    \item{\code{at}}{length-2 list of x/y points at which tickmarks are to be drawn, default \code{list(x=NA, y=NA)}}
    \item{\code{axes.lty}}{length-2 type of axis line (depends on \code{nx} value, c(0,1) if vertical gridlines are drawn) but can be modified here}
    \item{\code{gap.axis}}{length-2 minimal gap between axis labels}
    \item{\code{grid.col}}{gridline color}
    \item{\code{grid.lty}}{gridline type}
    \item{\code{grid.lwd}}{gridline width}
    \item{\code{main}}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}
    \item{\code{sub}}{Sub-title (at bottom) using font, size and color 
    \code{par(c("font.sub", "cex.sub", "col.sub"))}.}
    \item{\code{xlab}}{X axis label using font, size and color 
    \code{par(c("font.lab", "cex.lab", "col.lab"))}.}
    \item{\code{ylab}}{Y axis label, same font attributes as \code{xlab}.}
    \item{\code{line}}{specifying a value for \code{line} overrides the
    default placement of labels, and places them this many lines
    outwards from the plot edge.}
    \item{\code{nx,ny}}{number of cells of the grid in x and y direction.  When
    \code{NULL}, as per default, the grid aligns with the tick marks on
    the corresponding \emph{default} axis (i.e., tickmarks as computed by
    \code{\link[graphics]{axTicks}}).  When \code{\link{NA}}, no grid lines are
    drawn in the corresponding direction.}
    \item{\code{x,y}}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}
  }}

\item{side}{position of axes, up to 4 (1-bottom, 2-left, 3-top, 4-right)}

\item{use.brand}{whether to use mofidied plot layout, default \code{TRUE}}
}
\description{
Simply assign new default arguments to base R \code{plot} method. This function is exported but not meant to be called directly in most cases. Use \code{\link[=brand_on]{brand_on()}} instead to apply plot customizations.
}
\details{
Note that when package option \code{getOption(use.brand=FALSE)} then \code{plot.brand()} is entirely equivalent to \code{base::plot()}.
}
\examples{
set.seed(1)
x <- runif(100, min = -5, max = 5)
y <- x ^ 3 + rnorm(100, mean = 0, sd = 5)

# Standard plot
options(use.brand = FALSE)
plot.brand(x, y, col=4,
  main="Bootstrap Branded Plot", sub="Scatter plot")

# Branded defaults
options(use.brand = TRUE)
plot.brand(x, y, col=4,
  main="Bootstrap Branded Plot", sub="Scatter plot with moded axes")

plot.brand(x, y, type="h", col=(y>0)+4, side=c(1,4), nx=NULL,
  main="Bootstrap Branded Plot", sub="Histogram with dummy legend")
abline(h=0, col=pal.brand("red"), lwd=2)
legend.brand(names(pal.brand())[4:5], lty=1, lwd=2, col=4:5)

}
\seealso{
\code{\link[=axes]{axes()}}
}
