% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_brand}
\alias{plot_brand}
\title{Wrapper for base plot()}
\usage{
plot_brand(
  x,
  side = getOption("axes.side"),
  main = NA,
  sub = NA,
  main.line = 5.5,
  axes = FALSE,
  ann = axes,
  ...
)
}
\arguments{
\item{side}{which ide(s) to plot axes (0 to supress all axes), default to \code{getOption("axes.side")} else \code{c(1, 2)} (left and bottom)}

\item{main}{a main title for the plot, see also \code{\link[graphics]{title}}.}

\item{sub}{a subtitle for the plot.}

\item{main.line}{margin line to draw plot title and subtitle, used to adjust plot margins (default: 5.5)}

\item{axes}{a logical value indicating whether both axes should be drawn on
    the plot.  Use \link[graphics]{graphical parameter} \code{"xaxt"} or \code{"yaxt"}
    to suppress just one of the axes.}

\item{ann}{a logical value indicating whether the default annotation
    (title and x and y axis labels) should appear on the plot.}

\item{...}{
  Arguments passed on to \code{\link[graphics:plot.default]{graphics::plot.default}}
  \describe{
    \item{\code{x,y}}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}
    \item{\code{type}}{1-character string giving the type of plot desired.  The
    following values are possible, for details, see \code{\link[base]{plot}}:
    \code{"p"} for points,
    \code{"l"} for lines,
    \code{"b"} for both points and lines,
    \code{"c"} for empty points joined by lines,
    \code{"o"} for overplotted points and lines,
    \code{"s"} and \code{"S"} for stair steps and
    \code{"h"} for histogram-like vertical lines.  Finally,
    \code{"n"} does not produce any points or lines.}
    \item{\code{xlim}}{the x limits (x1, x2) of the plot.  Note that \code{x1 > x2}
    is allowed and leads to a \sQuote{reversed axis}.

    The default value, \code{NULL}, indicates that the range of the
    \link{finite} values to be plotted should be used.}
    \item{\code{ylim}}{the y limits of the plot; see \code{xlim} for default and details.}
    \item{\code{log}}{a character string which contains \code{"x"} if the x axis
    is to be logarithmic, \code{"y"} if the y axis is to be logarithmic
    and \code{"xy"} or \code{"yx"} if both axes are to be logarithmic.}
    \item{\code{lim2}}{logical, relevant only when at least one of \code{xlim} or
    \code{ylim} are \code{NULL} and hence corresponding axis limits are
    computed from \code{\link{range}(<x/y-coord>[<finite-coord>])}.
    When true, \emph{both} x- and y-values need to be finite to be used in
    the \code{range(.)} computation.}
    \item{\code{xlab}}{a label for the x axis, defaults to a description of \code{x}.}
    \item{\code{ylab}}{a label for the y axis, defaults to a description of \code{y}.}
    \item{\code{frame.plot}}{a logical indicating whether a box should be drawn
    around the plot.}
    \item{\code{panel.first}}{an \sQuote{expression} to be evaluated after the
    plot axes are set up but before any plotting takes place.  This can
    be useful for drawing background grids or scatterplot smooths.  Note
    that this works by lazy evaluation: passing this argument from other
    \code{plot} methods may well not work since it may be evaluated too
    early.}
    \item{\code{panel.last}}{an expression to be evaluated after plotting has
    taken place but before the axes, title and box are added.  See the
    comments about \code{panel.first}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio, see \code{\link[graphics]{plot.window}}.}
    \item{\code{xgap.axis,ygap.axis}}{the \eqn{x/y} axis gap factors, passed as
    \code{gap.axis} to the two \code{\link[graphics]{axis}()} calls (when
    \code{axes} is true, as per default).}
  }}
}
\description{
Simply assign default arguments to base R \code{plot} method. This function is exported but not meant to be called directly in most cases. Use \code{\link[=brand_on]{brand_on()}} to apply complete plot customizations (fonts and colors) using \verb{_brand.yml} features.
}
\details{
Use \code{side = c(1,2)} etc. to control axis visibility. For more granular control over axis display use \code{side=0} and this package \code{\link[=axes]{axes()}} or base R \code{\link[=axis]{axis()}}.
}
\examples{
set.seed(1)
x <- runif(100, min = -5, max = 5)
y <- x ^ 3 + rnorm(100, mean = 0, sd = 5)

# Standard plot
plot_brand(x, y, col=4, side=1:2, main.line=2.5,
  main="Custom Plot", sub="no legend", axes=TRUE, ann=TRUE)

# Standard plot
plot_brand(x, y, col=4, side=0, main.line=2.5,
  main="Custom Plot", sub="no axis")

# Branded defaults
plot_brand(x, y, type="h", col=(y>0)+4, side=c(1,4), nx=NULL,
  main="Custom Plot", sub="Histogram with top legend")
abline(h=0, col=pal("red"), lwd=2)
legend_brand(names(pal())[4:5], lty=1, lwd=2, col=4:5)

}
\seealso{
\code{\link[=axes]{axes()}}
}
