% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tibble.R
\name{ptbl}
\alias{ptbl}
\title{convenience function to print tibbles.}
\usage{
ptbl(x, n = NULL, ...)
}
\arguments{
\item{x}{a tibble to be printed}

\item{n}{number of rows to be printed, or all (unquoted).}

\item{...}{arguments to be passed to print.tbl_df}
}
\description{
convenience function to print tibbles.
}
\details{
first, a default max rows is determined. Pretty much a wrapper to the standard tibble print function, but
  defaults to all rows and columns (but with a maximum).
}
