% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode-pattern.R
\docType{data}
\name{rcode_pattern}
\alias{rcode_pattern}
\alias{Rcode_pattern}
\alias{pat_rcode}
\alias{pat_rlatex}
\alias{pat_Rcode}
\title{Definition of the rcode pattern}
\format{An object of class \code{list} of length 7.}
\usage{
Rcode_pattern

pat_rcode()

pat_rlatex()

pat_Rcode()
}
\description{
Definition of the rcode pattern

Functions to set the rcode pattern before knitting or rendering.
}
\details{
A named list defining the rcode pattern.

Sets the pattern to the Rcode default pattern, using the
  same mechanism as knitr. When you run this, you will not be able
  to knit or render other documents without setting the pattern, or
  clearing using knit_patterns$restore().
This will work with Rnw files.  So, to compile an Rnw file with
  Rcode blocks, run pat_rcode() first.  To switch back to Sweave
  syntax (or really, any other syntax), run
  knit_patterns#restore(), and let knitr figure it out. Or, run
  pat_zzz(), where zzz is an extension for the various convenience
  functions provided by knitr.
}
\keyword{datasets}
