% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode-pattern.R
\docType{data}
\name{rmd_pattern}
\alias{rmd_pattern}
\alias{Rmd_pattern}
\alias{pat_rmd}
\alias{pat_Rmd}
\title{Definition of the Rmd pattern}
\format{
An object of class \code{list} of length 4.
}
\usage{
Rmd_pattern

pat_rmd()

pat_Rmd()
}
\description{
Definition of the Rmd pattern
}
\details{
A named list defining the rmd pattern
}
\keyword{datasets}
