% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqtab}
\alias{seqtab}
\title{Example sequence table}
\format{
## `seqtab` A data frame with 52 rows and 110 columns:
\describe{
  \item{ATCACGTTTTCCGAAAACAAACAACGGTTCAGAAAGCGAAAATCAAAAAG}{ASV1}
  \item{ATCCCTTTTTTGAAAAACAAGTGGTTCTCAAACTAGAACCCAAAGGAAAAG}{ASV2}
  ...
}
}
\usage{
seqtab
}
\description{
Rows are samples (named here by their plate and well location), and columns
are *trnL* amplicon sequence variants (ASVs). Values indicate counts of an
ASV detected in a given sample inferred by DADA2.
}
\keyword{datasets}
