% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.balance.R
\name{time.balance}
\alias{time.balance}
\title{Obtain time balance comparing expected time and employed time in a process}
\usage{
\method{time}{balance}(expected.time, employed.time, justified.time = 0)
}
\arguments{
\item{expected.time}{A string. A vector with the expected time to reach
process achievement. It can be the sum of the total expected time or the
partial times that sum up to the total expected time. Format: "hh:mm".}

\item{employed.time}{A string. A vector with the time employed to reach
process achievement. It can be the sum of the total employed time or the
partial times that sum up to the total employed time. Format: "hh:mm".}

\item{justified.time}{A string. A vector with the total time not directly
employed in the process but justified as part of the necessary time to
achieve the process. It can be the sum of the total justified time or the
partial times that sum up to the total justified time. Format: "hh:mm".}
}
\value{
time balance between expected time and employed  + justified time.
}
\description{
Obtain time balance comparing expected time and employed time in a process
}
\examples{
time.balance(expected.time = "8:00", employed.time = "5:30",
justified.time = "2:00")

}
\author{
Manuel Molina-Bustamante
}
