% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_assignment.R
\name{update_assignment}
\alias{update_assignment}
\title{Update the taxonomic assignment of a phyloseq object}
\usage{
update_assignment(ps, ref)
}
\arguments{
\item{ps}{Phyloseq object to be updated}

\item{ref}{Path to the reference database (a FASTA file) to use for
assignment}
}
\value{
An updated phyloseq object.
}
\description{
Replaces the tax_table slot of a phyloseq object with updated
  assignments based on a revised version of a reference database.
}
