% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{brand.colors}
\alias{brand.colors}
\title{Bootstrap color ramp}
\usage{
brand.colors(x = NULL, omit = c("white", "black", "gray", "grid"), ...)
}
\arguments{
\item{x}{Brand color name(s), skip to return the entire color palette}

\item{omit}{Brand colors to exclude from color ramp}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette}}
  \describe{
    \item{\code{colors}}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}
    \item{\code{bias}}{a positive number.  Higher values give more widely spaced
    colors at the high end.}
    \item{\code{space}}{a character string; interpolation in RGB or \abbr{CIE} Lab
    color spaces.}
    \item{\code{interpolate}}{use spline or linear interpolation.}
    \item{\code{alpha}}{logical: should alpha channel (opacity) values be
    returned?   It is an error to give a true value if
    \code{space} is specified.}
  }}
}
\value{
color palette function
}
\description{
Qualitative color ramp derived from active branding. This ramp excludes
Bootstrap's \strong{white}, \strong{black}, \strong{grid} and \strong{gray} colors, which are
typically used for textual elements.
}
\examples{
scales::show_col(brand.colors()(6))
scales::show_col(brand.colors(interpolate="spline")(12))
scales::show_col(brand.colors(c("orange", "light"))(12))

}
