% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{brand.colors}
\alias{brand.colors}
\title{Bootstrap color ramp}
\usage{
brand.colors(x = NULL, omit = c("white", "black", "gray"), ...)
}
\arguments{
\item{x}{Brand color name(s), skip to return the entire color palette}

\item{omit}{Brand colors to exclude from color ramp}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette}}
  \describe{
    \item{\code{colors}}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}
  }}
}
\value{
vector of n interpolated colors
}
\description{
Qualitative color ramp derived from active branding. This ramp excludes Bootstrap's
\strong{white}, \strong{black}, \strong{light} and \strong{gray} colors, which are typically used for
textual elements. By default the color ramp is 90\% transparent.
}
\examples{
scales::show_col(brand.colors()(6))
scales::show_col(brand.colors(interpolate="spline")(12))
scales::show_col(brand.colors(c("orange", "light"))(12))
s
}
