% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{brand}
\alias{brand}
\title{Fetch Bootstrap branding}
\usage{
brand(file = "_brand.yml", font = c("base", "monospace", "headings"))
}
\arguments{
\item{file}{path to \verb{_brand.yml} configuration file, normally this file is auto-detected in the working tree, but may be specified here to swap branding dynamically.}

\item{font}{one of \verb{_brand.yml} font families (currently only \code{base}, \code{monospace}, or \code{headings}).}
}
\value{
List of branding elements similar in structure to \verb{_brand.yml}
}
\description{
Make this package compatible with \code{brand.yml} unified branding feature. Branding can be read from a configuration file provided at run-time, else it will use one of these package built-in themes.
}
\details{
Refer to \href{https://posit-dev.github.io/brand-yml/}{brand.yml} documentation.

Typically there is no need to call \code{brand()} directly. Instead use \code{\link[=brand_on]{brand_on()}} to load and apply branding elements to your active session.

TODO search more file locations
}
\examples{
brand(font="monospace")
scales::show_col(unlist(brand()$color$palette))

}
\references{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
\seealso{
\code{\link[=brand_on]{brand_on()}}
}
