% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{date.annotation}
\alias{date.annotation}
\title{obtain annotation of datetime as string.}
\usage{
date.annotation(date = Sys.time())
}
\arguments{
\item{date}{"POSIXct" "POSIXt". Default Sys.time().}
}
\value{
character with datime
}
\description{
Obtain annotation of datetime as string.
}
\examples{
date.annotation(date = "2022-12-14 13:09:53 CET")
date.annotation()

}
\author{
Manuel Molina-Bustamante
}
