% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directorytree.R
\name{directorytree}
\alias{directorytree}
\title{Create a Directory Tree}
\usage{
directorytree(path, only_directories = TRUE)
}
\arguments{
\item{path}{A character string specifying the root directory path.}

\item{only_directories}{A logical value indicating whether to include only
directories (TRUE) or both files and directories (FALSE). Default is TRUE.}
}
\value{
A Node object representing the directory tree.
}
\description{
This function creates a directory tree structure starting from a specified
path.
}
\examples{
\dontrun{
tree <- directorytree("/path/to/directory", only_directories = TRUE)
print(tree)
}
}
