% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_level.R
\name{lowest_level}
\alias{lowest_level}
\title{Extract the finest taxonomic level of ASVs from a phyloseq taxonomy}
\usage{
lowest_level(taxtab, unique = FALSE)
}
\arguments{
\item{taxtab}{The taxonomy table of the phyloseq object to be renamed}

\item{unique}{If TRUE, appends an integer count to taxa names that are
duplicated across ASVs.}
}
\value{
An updated taxonomy table, with an added column, 'name', that
  contains the name of the lowest phylogenetic level to which that ASV is
  identified.  If unique = TRUE, duplicates are labeled uniquely by appending
  an integer.
}
\description{
Identifies the lowest taxonomic name available for each ASV in a
  phyloseq taxonomy table and aggregates them in a new column.
}
