% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{pal}
\alias{pal}
\title{Bootstrap color palette}
\usage{
pal(x = NULL, alpha = 0.85, named = TRUE)
}
\arguments{
\item{x}{Color index or name(s), skip to return the entire color palette}

\item{alpha}{transparency (default: 0.85)}

\item{named}{keep color names (default: TRUE)}
}
\value{
A vector of (named) hex color codes extracted from Bootstrap branding
}
\description{
Color palette extracted from the active Bootstrap theme. By default colors are read from an external \verb{_brand.yml} configuration file (or uses this package built-in brand).
}
\examples{
scales::show_col(pal())
scales::show_col(pal(c("orange", "red")))

}
\references{
\href{https://posit-dev.github.io/brand-yml/}{brand.yml}
}
