% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Example taxonomy data structure}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 10 columns.
}
\usage{
taxonomy
}
\description{
Example taxonomy data structure
}
\keyword{datasets}
